Quickfix for for circular dependency between AceConfig and Pattern
